Refactor entity IDs to use long type, update related tests, and add appsettings for Snowflake configuration

+++ src/Domain/Common/Entity.cs
@@ -2,7 +2,7 @@ namespace DeveloperStore.Domain.Common;
 
 public abstract class Entity
 {
-    public int Id { get; protected set; }
+    public long Id { get; protected set; }
 
     public override bool Equals(object obj)
     {
