+++ src/Domain/Events/ItemCancelledEvent.cs
@@ -6,10 +6,10 @@ namespace DeveloperStore.Domain.Events;
 public class ItemCancelledEvent : IDomainEvent
     {
         public Sale Sale { get; }
-        public int ProductId { get; }
+        public long ProductId { get; }
         public DateTime OccurredOn { get; }
 
-        public ItemCancelledEvent(Sale sale, int productId)
+        public ItemCancelledEvent(Sale sale, long productId)
         {
             Sale = sale;
             ProductId = productId;
